+++ webpack.config.js
@@ -1,54 +1,53 @@
-const path = require("path");
-const webpack = require("webpack");
-const HtmlWebpackPlugin = require("html-webpack-plugin");
-const UglifyJsPlugin = require("uglifyjs-webpack-plugin");
-const { CleanWebpackPlugin } = require("clean-webpack-plugin");
+const path = require('path');
+const webpack = require('webpack');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
+const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 // const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin
-const utils = require("./build/utils");
-const theme = require("./build/theme");
-const HtmlWebpackInlineSourcePlugin = require("./build/InlineChunkHtmlPlugin");
+const utils = require('./build/utils');
+const theme = require('./build/theme');
 
-const packageInfo = require("./package.json");
+const packageInfo = require('./package.json');
 function resolve(dir) {
   return path.join(__dirname, dir);
 }
 
 module.exports = {
-  mode: "production",
+  mode: 'production',
   entry: {
-    index: "./src/index.js",
+    index: './src/index.js',
   },
   output: {
-    path: resolve("dist"),
-    filename: "[name].js",
+    path: resolve('dist'),
+    publicPath: './jest-html-reporters-attach',
+    filename: '[name].js',
   },
   resolve: {
-    extensions: [".js", ".json"],
+    extensions: ['.js', '.json'],
     alias: {
-      "@": resolve("src"),
+      '@': resolve('src'),
     },
   },
   plugins: [
     new CleanWebpackPlugin(),
     new webpack.DefinePlugin({
-      "process.env": {
-        NODE_ENV: JSON.stringify("production"),
+      'process.env': {
+        NODE_ENV: JSON.stringify('production'),
         VERSIONS: JSON.stringify(packageInfo.version),
       },
     }),
     new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
-    ...["index"].map(
+    ...['index'].map(
       (name) =>
         new HtmlWebpackPlugin({
           filename: `${name}.html`,
           chunks: [name],
-          template: "index.html",
-          inlineSource: ".(js|css)$",
+          template: 'index.html',
+          inlineSource: '.(js|css)$',
         })
     ),
-    new HtmlWebpackInlineSourcePlugin(HtmlWebpackPlugin, [/s/]),
     // Copyright
-    new webpack.BannerPlugin("Copyright Harry All rights reserved."),
+    new webpack.BannerPlugin('Copyright Harry All rights reserved.'),
     // new BundleAnalyzerPlugin(),
   ],
   optimization: {
@@ -68,20 +67,20 @@ module.exports = {
     rules: [
       {
         test: /\.js$/,
-        loader: "babel-loader",
-        include: [resolve("src"), resolve("test")],
+        loader: 'babel-loader',
+        include: [resolve('src'), resolve('test')],
       },
       {
         test: /\.css$/,
-        use: ["style-loader", "css-loader"],
+        use: ['style-loader', 'css-loader'],
       },
       {
         test: /\.less$/,
         use: [
-          "style-loader",
-          "css-loader",
+          'style-loader',
+          'css-loader',
           {
-            loader: "less-loader",
+            loader: 'less-loader',
             options: {
               lessOptions: {
                 javascriptEnabled: true,
@@ -93,36 +92,36 @@ module.exports = {
       },
       {
         test: /\.(png|jpe?g|gif)(\?.*)?$/,
-        loader: "url-loader",
+        loader: 'url-loader',
         options: {
           limit: 10000,
-          name: utils.assetsPath("img/[name].[hash:7].[ext]"),
+          name: utils.assetsPath('img/[name].[hash:7].[ext]'),
         },
       },
       {
         test: /\.(mp4|webm|ogg|mp3|wav|flac|aac)(\?.*)?$/,
-        loader: "url-loader",
+        loader: 'url-loader',
         options: {
           limit: 10000,
-          name: utils.assetsPath("media/[name].[hash:7].[ext]"),
+          name: utils.assetsPath('media/[name].[hash:7].[ext]'),
         },
       },
       {
         test: /\.(woff2?|eot|ttf|otf)(\?.*)?$/,
-        loader: "url-loader",
+        loader: 'url-loader',
         options: {
           limit: 10000,
-          name: utils.assetsPath("fonts/[name].[hash:7].[ext]"),
+          name: utils.assetsPath('fonts/[name].[hash:7].[ext]'),
         },
       },
       {
         test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
         use: [
           {
-            loader: "babel-loader",
+            loader: 'babel-loader',
           },
           {
-            loader: "@svgr/webpack",
+            loader: '@svgr/webpack',
             options: {
               babel: false,
               icon: true,